+++ src/main/java/com/salesforce/dynamodbv2/mt/mappers/sharedtable/impl/MtAmazonDynamoDbStreamsBySharedTable.java
@@ -112,7 +112,7 @@ public class MtAmazonDynamoDbStreamsBySharedTable extends MtAmazonDynamoDbStream
                 .withRecords(new ArrayList<>(limit))
                 .withNextShardIterator(request.getShardIterator());
 
-            final RecordMapper recordMapper =
+            final IRecordMapper recordMapper =
                 mtDynamoDb.getTableMapping(mtStreamArn.getTenantTableName()).getRecordMapper();
             final Predicate<Record> recordFilter = recordMapper.createFilter();
